tools/buttonboard/companion/requirements.txt
flask
flask-cors
pyautogui
pystray
pillow
bless
